# Optimize LocalMerge operator