- Remove extra copying into RowContainer.
- Use TreeOfLosers for merging.
- Copy rows to output in bulk.

This version is much faster than previous. Merging 7 streams of 80M rows each on
1 integer key is 27s vs original 70s and on 2 keys is 32s vs original 290s.
These numbers include optimizations for FlatVector::copyValuesAndNulls
(#1316) and BaseVector::compare (#1317)

1 flat int key
<img width="1047" alt="Screen Shot 2022-03-31 at 7 03 18 AM" src="https://user-images.githubusercontent.com/27965151/161041090-1750a8c0-9f82-4bb1-9e96-5bc2d4929691.png">

2 keys: flat and const:

<img width="1044" alt="Screen Shot 2022-03-31 at 7 03 44 AM" src="https://user-images.githubusercontent.com/27965151/161041104-c38a923d-96dc-4f05-bb2f-a42c2f409b62.png">